Change how kortvaerk works
Now returns the specific kortvaerk client have requested with an IN search instead of SIMILAR TO.

We can't use List<String> to kortvaerks type in the query param, because Spring Boot does not respect the url encoded `,` as `%2C` if a kortvaerk has comma in its name, it splits it appart because of the `,` even if it is a valid name.
So we use this workaround to make it possible.
https://www.github.com/spring-projects/spring-framework/issues/23820#issuecomment-1059578991